Add changeActiveSlas and selectSlaFromSlaOption functions
