+++ react/modules.js
@@ -2,10 +2,24 @@ import {
   getLeanShippingOptions,
   getOptionsDetails,
   getSelectedDeliveryOption,
-} from './index'
+} from './leanShipping'
+import {
+  setSelectedSlaFromSlaOption,
+  changeActiveSlas,
+} from './changeActiveSlas'
+import {
+  setSelectedDeliveryChannel,
+  findPickupSla,
+  findFirstItemWithPickup,
+} from './utils'
 
 export default {
   getLeanShippingOptions,
+  setSelectedSlaFromSlaOption,
+  setSelectedDeliveryChannel,
   getOptionsDetails,
   getSelectedDeliveryOption,
+  findPickupSla,
+  changeActiveSlas,
+  findFirstItemWithPickup,
 }
