+++ react/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@vtex/lean-shipping-calculator",
   "dependencies": {
+    "@vtex/delivery-packages": "^2.17.0",
     "@vtex/estimate-calculator": "1.0.7",
     "lodash": "^4.17.5"
   }
