+++ tests/__mocks__/vtex.address-form.js
@@ -1,0 +1,5 @@
+const helpers = {
+  removeValidation: address => address,
+}
+
+module.exports = { helpers }
